+++ src/com/amazon/pay/api/ServiceConstants.java
@@ -26,7 +26,7 @@ public class ServiceConstants {
     public static final Map<Region, String> endpointMappings;
     public static final Map<String, Integer> serviceErrors;
 
-    public static final String APPLICATION_LIBRARY_VERSION = "2.6.6";
+    public static final String APPLICATION_LIBRARY_VERSION = "2.6.7";
     public static final String GITHUB_SDK_NAME = "amazon-pay-api-sdk-java";
     public static final String AMAZON_PAY_API_VERSION = "v2";
 
@@ -88,6 +88,7 @@ public class ServiceConstants {
         Map<String, Integer> serviceErrorsMap = new HashMap<>();
         serviceErrorsMap.put("Internal Server Error", HttpURLConnection.HTTP_INTERNAL_ERROR);
         serviceErrorsMap.put("Service Unavailable", HttpURLConnection.HTTP_UNAVAILABLE);
+        serviceErrorsMap.put("Too Early", 425);
         serviceErrorsMap.put("Too Many Requests", 429);
         serviceErrorsMap.put("HTTP Bad Gateway", HttpURLConnection.HTTP_BAD_GATEWAY);
         serviceErrorsMap.put("HTTP Gateway Timeout", HttpURLConnection.HTTP_GATEWAY_TIMEOUT);